pyproject.toml
[tool.poetry]
name = "dfolks"
version = "0.1.0"
description = "Data analytics platform developed by dfolks"
authors = ["dfolks <digitalfolks55@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "dfolks", from = "src"}
]


[tool.poetry.dependencies]
python = "~3.12"
pandas = "~2.2.3"
numpy = "~2.2.3"
phx-class-registry = "^5.1.1"
pydantic = "^2.10.6"
pydantic-yaml = "^1.4.0"
pyyaml = "^6.0.2"
pandera = "^0.23.1"
ipykernel = "^6.29.5"
openpyxl = "^3.1.5"
scikit-learn = "^1.6.1"
class-registry = "^2.1.2"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
python-xbrl = "^1.1.1"
beautifulsoup4 = "^4.13.4"
arelle-release = "^2.37.60"
google-api-python-client = "^2.186.0"
google-auth-httplib2 = "^0.2.1"
google-auth-oauthlib = "^1.2.3"


[tool.poetry.group.dev.dependencies]
black = "25.1.0"
isort = "6.0.1"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pre-commit = "^4.1.0"
flake8 = "7.1.2"


[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
